refactor: 에러 핸들링 헬퍼 함수 추가
