+++ src/shared/utils/errorHandling.ts
@@ -1,0 +1,3 @@
+export const handleError = (error: unknown) => {
+	if (error) throw error;
+};
